+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Changelog
 
+- 1.0.6: FEAT: Kick off neglected leetcoders after 5 non-solved problems
+- 1.0.5: FEAT: Secure cron job
+- 1.0.4: ENHANCEMENT: Use [cron-jon](https://console.cron-job.org/jobs) instead of vercel cron
 - 1.0.3: FEAT: Automate "send daily problem email" cron job
 - 1.0.2: FEAT: Automate "add group progress" cron job
 - 1.0.1: FIX: Show validation error message when the form is submitted